Making SSL mount optional

+++ Web-Ods-AdminApp/Alpine/pgsql/run.sh
@@ -40,8 +40,9 @@ done
 >&2 echo "Postgres is up - executing command"
 exec $cmd
 
-
-cp /ssl/server.crt /usr/local/share/ca-certificates/
-update-ca-certificates
+if [ -f /ssl/server.crt ]; then
+  cp /ssl/server.crt /usr/local/share/ca-certificates/
+  update-ca-certificates
+fi
 
 dotnet EdFi.Ods.AdminApp.Web.dll
